Improved on UX
Only register form is shown on initial page load. sign in, update,
signout are initially hidden. Toggle link below form toggles sign in
form. On sign in success update-password and sign-out forms are
revealed. on Sign Out sign out update and sign out are hidden and
sign-in form is shown.