bug fix on calculateStats conditional.
calculate states is future-proofed to check which token the user is
playing as, o or x. Since the api defaults as x, there will always be an
ID at player_x. the if statement for checking whether the user was
playing as o assumed there was an id in player_o as well, though a null
is most often the case.

solution for bug is to check if winning token is o and player_o != null.
then check player o id to the user's id.