resume last game functionality complete
Finished onResumeGame function that takes the last object in the array
returned by GET /games?over=false request and makes it the current game.
Also added functions that updated the boardUI according to what values
board.cells has.

Note: the new ui.resumeLastGame function can be used to update board
when user clicks a valid cell. Should compare with current solution in
the future and see which is better.